refactor global state handling
