+++ handlers/client.go
@@ -4,7 +4,7 @@ type OutboundHandler struct {
 	*Handler
 }
 
-func NewOutboundHandler(apikey, clientId, environment, service string, otelEnabled, sentinelEnabled bool) (*OutboundHandler, error) {
-	h, err := NewHandler(apikey, clientId, environment, service, otelEnabled, sentinelEnabled)
+func NewOutboundHandler() (*OutboundHandler, error) {
+	h, err := NewHandler()
 	return &OutboundHandler{h}, err
 }
