+++ handlers/handler.go
@@ -1,6 +1,7 @@
 package handlers
 
 import (
+	"github.com/StanzaSystems/sdk-go/global"
 	"github.com/StanzaSystems/sdk-go/otel"
 	hubv1 "github.com/StanzaSystems/sdk-go/proto/stanza/hub/v1"
 	"go.opentelemetry.io/otel/attribute"
@@ -9,79 +10,54 @@ import (
 )
 
 type Handler struct {
-	apikey          string
-	clientId        string
-	customerId      string
-	environment     string
-	otelEnabled     bool
-	sentinelEnabled bool
-
-	decoratorConfig map[string]*hubv1.DecoratorConfig
-	qsc             hubv1.QuotaServiceClient
-	propagators     propagation.TextMapPropagator
-	tracer          trace.Tracer
-	meter           *Meter
-	attr            []attribute.KeyValue
+	tracer trace.Tracer
+	meter  *Meter
+	attr   []attribute.KeyValue
 }
 
-func NewHandler(apikey, clientId, environment, service string, otelEnabled, sentinelEnabled bool) (*Handler, error) {
+func NewHandler() (*Handler, error) {
 	m, err := GetStanzaMeter()
 	return &Handler{
-		apikey:          apikey,
-		clientId:        clientId,
-		environment:     environment,
-		otelEnabled:     otelEnabled,
-		sentinelEnabled: sentinelEnabled,
-		decoratorConfig: make(map[string]*hubv1.DecoratorConfig),
-		qsc:             nil,
-		propagators:     otel.GetTextMapPropagator(),
-		meter:           m,
+		meter: m,
 		tracer: otel.GetTracerProvider().Tracer(
 			instrumentationName,
 			trace.WithInstrumentationVersion(instrumentationVersion),
 		),
 		attr: []attribute.KeyValue{
-			clientIdKey.String(clientId),
-			environmentKey.String(environment),
-			serviceKey.String(service),
+			clientIdKey.String(global.GetClientID()),
+			environmentKey.String(global.GetServiceEnvironment()),
+			serviceKey.String(global.GetServiceName()),
 		},
 	}, err
-}
-
-func (h *Handler) APIKey() string {
-	return h.apikey
-}
-
-func (h *Handler) Attributes() []attribute.KeyValue {
-	return h.attr
-}
-
-func (h *Handler) ClientID() string {
-	return h.clientId
-}
-
-func (h *Handler) DecoratorConfig(dec string) *hubv1.DecoratorConfig {
-	return h.decoratorConfig[dec]
-}
-
-func (h *Handler) DecoratorKey(dec string) attribute.KeyValue {
-	return decoratorKey.String(dec)
-}
-
-func (h *Handler) Environment() string {
-	return h.environment
-}
-
-func (h *Handler) FeatureKey(feat string) attribute.KeyValue {
-	return featureKey.String(feat)
 }
 
 func (h *Handler) Meter() *Meter {
 	return h.meter
 }
 
+func (h *Handler) Tracer() trace.Tracer {
+	return h.tracer
+}
+
 func (h *Handler) Propagator() propagation.TextMapPropagator {
 	return otel.GetTextMapPropagator()
+}
+
+// OTEL Attribute Helper Functions //
+func (h *Handler) Attributes() []attribute.KeyValue {
+	return h.attr
+}
+
+func (h *Handler) CustomerKey() attribute.KeyValue {
+	return customerIdKey.String(global.GetCustomerID())
+}
+
+func (h *Handler) DecoratorKey(dec string) attribute.KeyValue {
+	return decoratorKey.String(dec)
+}
+
+func (h *Handler) FeatureKey(feat string) attribute.KeyValue {
+	return featureKey.String(feat)
 }
 
 func (h *Handler) ReasonKey(reason string) attribute.KeyValue {
@@ -100,33 +76,31 @@ func (h *Handler) ReasonQuota() attribute.KeyValue {
 	return reasonKey.String("quota")
 }
 
-func (h *Handler) SetCustomerId(id string) {
-	if h.customerId == "" {
-		h.customerId = id
-		h.attr = append(h.attr, customerIdKey.String(id))
-	}
+// Global Helper Functions //
+func (h *Handler) APIKey() string {
+	return global.GetServiceKey()
 }
 
-func (h *Handler) SetDecoratorConfig(d string, dc *hubv1.DecoratorConfig) {
-	if h.decoratorConfig[d] == nil {
-		h.decoratorConfig[d] = dc
-	}
+func (h *Handler) ClientID() string {
+	return global.GetClientID()
 }
 
-func (h *Handler) SetQuotaServiceClient(quotaServiceClient hubv1.QuotaServiceClient) {
-	if h.qsc == nil {
-		h.qsc = quotaServiceClient
-	}
+func (h *Handler) DecoratorConfig(decorator string) *hubv1.DecoratorConfig {
+	return global.DecoratorConfig(decorator)
+}
+
+func (h *Handler) Environment() string {
+	return global.GetServiceEnvironment()
+}
+
+func (h *Handler) OTELEnabled() bool {
+	return global.OtelEnabled()
 }
 
 func (h *Handler) SentinelEnabled() bool {
-	return h.sentinelEnabled
-}
-
-func (h *Handler) Tracer() trace.Tracer {
-	return h.tracer
+	return global.SentinelEnabled()
 }
 
 func (h *Handler) QuotaServiceClient() hubv1.QuotaServiceClient {
-	return h.qsc
+	return global.QuotaServiceClient()
 }
